added the authentication protect for question and answer apis

+++ application/controllers/Answers.php
@@ -40,31 +40,42 @@
         // create answer
         public function create_post() {
             try {
-                $question_id = $this->post('question_id');
-                $body = $this->post('body');
-                // might need to fetch the user id from session
-                // will be implemented when the user login feature completed
-                $user_id = '25614802-8592-11ed-a1eb-0242ac120002';
-                if($question_id === "" || $body === "" || $user_id === "" ) {
+                // user authentication
+                // validate the user
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                $decodedToken = (array)decodeJWTToken($token);
+                if ($isValid && $this->session->userdata('logged_in')) {
+                    $question_id = $this->post('question_id');
+                    $body = $this->post('body');
+                    $user_id = $decodedToken['user_id'];
+                    if($question_id === "" || $body === "" || $user_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters are missing',
+                            'result' => null,
+                        ], 400);
+                    }
+                    $data = array(
+                        'answer_id' => Uuid::uuid1(),
+                        'question_id' => $question_id,
+                        'user_id' => $user_id,
+                        'body' => $body
+                    );
+
+                    $result = $this->answer_model->create_answer($data);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 201);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters are missing',
-                        'result' => null,
-                    ], 400);
+                        'error' => 'unauthorized',
+                        'result' => null
+                    ], 401);
                 }
-                $data = array(
-                    'answer_id' => Uuid::uuid1(),
-                    'question_id' => $question_id,
-                    'user_id' => $user_id,
-                    'body' => $body
-                );
-
-                $result = $this->answer_model->create_answer($data);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 201);
             } catch (Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -78,25 +89,38 @@
         // update answer
         public function update_put($answer_id) {
             try {
-                $body = $this->put('body');
-                if ($body === "" || $answer_id === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                $decodedToken = (array)decodeJWTToken($token);
+                $user_id = $this->answer_model->get_user($answer_id);
+                // only created user can edit the question
+                if ($isValid && $this->session->userdata('logged_in') && $user_id === $decodedToken['user_id']) {
+                    $body = $this->put('body');
+                    if ($body === "" || $answer_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters missing',
+                            'result' => null
+                        ], 400);
+                    }
+    
+                    $data = array(
+                        'body' => $body
+                    );
+    
+                    $result = $this->answer_model->update_answer($answer_id, $data);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 200);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters missing',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
+                    ], 401);
                 }
-
-                $data = array(
-                    'body' => $body
-                );
-
-                $result = $this->answer_model->update_answer($answer_id, $data);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 200);
             } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -110,16 +134,28 @@
         // delete answer
         public function delete_delete($answer_id) {
             try {
-                if($answer_id === "") {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                $decodedToken = (array)decodeJWTToken($token);
+                $user_id = $this->answer_model->get_user($answer_id);
+                if ($isValid && $this->session->userdata('logged_in') && $user_id === $decodedToken['user_id']) {
+                    if($answer_id === "") {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters missing',
+                            'result' => null
+                        ], 400);
+                    } 
+        
+                    $result = $this->answer_model->delete_answer($answer_id);
+                    $this->response([], 204);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters missing',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
-                } 
-    
-                $result = $this->answer_model->delete_answer($answer_id);
-                $this->response([], 204);
+                    ], 401);
+                }
             } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -133,20 +169,30 @@
         // upvote for an answer
         public function upvote_put($answer_id) {
             try {
-                if ($answer_id === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                if ($isValid && $this->session->userdata('logged_id')) {
+                    if ($answer_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters missing',
+                            'result' => null,
+                        ], 400);
+                    }
+        
+                    $result = $this->answer_model->increase_vote($answer_id);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 200);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters missing',
-                        'result' => null,
-                    ], 400);
+                        'error' => 'unauthorized',
+                        'result' => null
+                    ], 401);
                 }
-    
-                $result = $this->answer_model->increase_vote($answer_id);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 200);
             } catch(Exception $e) {
                log_message('error', 'error =>'.$e->getMessage());
                $this->response([
@@ -160,20 +206,30 @@
         // downvote for an answer
         public function downvote_put($answer_id) {
            try {
-            if (!$answer_id) {
+            $token = $this->input->request_headers()['x-auth'];
+            $isValid = validateJWTToken($token);
+            if ($isValid && $this->session->userdata('logged_in')) {
+                if (!$answer_id) {
+                    $this->response([
+                        'code' => -1,
+                        'error' => 'mandatory parameters',
+                        'result' => null
+                    ], 400);
+                }
+    
+                $result = $this->answer_id->decrease_model($answer_id);
+                $this->response([
+                    'code' => 0,
+                    'error' => null,
+                    'result' => $result
+                ], 200);
+            } else {
                 $this->response([
                     'code' => -1,
-                    'error' => 'mandatory parameters',
+                    'error' => 'unauthorized',
                     'result' => null
-                ], 400);
+                ], 401);
             }
-
-            $result = $this->answer_id->decrease_model($answer_id);
-            $this->response([
-                'code' => 0,
-                'error' => null,
-                'result' => $result
-            ], 200);
            } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
