+++ application/controllers/Questions.php
@@ -131,33 +131,46 @@
         // update question
         public function update_put($question_id) {
             try {
-                $user_id = '25614802-8592-11ed-a1eb-0242ac120002';
-                $title = $this->put('title');
-                $body = $this->put('body');
-                if( $user_id === "" || $title === "" || $body === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                $decodedToken = (array)decodeJWTToken($token);
+                $user_id = $this->question_model->get_user($question_id);
+                // only created user can edit the question
+                if ($isValid && $this->session->userdata('logged_in') && $user_id === $decodedToken['user_id']) {
+                    // $user_id = '25614802-8592-11ed-a1eb-0242ac120002';
+                    $title = $this->put('title');
+                    $body = $this->put('body');
+                    if( $user_id === "" || $title === "" || $body === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters are empty',
+                            'result' => null
+                        ], 400);
+                    }
+    
+                    $timestamp = time();
+                    // 2022-12-27 11:12:02
+    
+                    $data = array(
+                        'title' => $title,
+                        'body' => $body,
+                        'updatedAt' => date("Y-m-d h:m:s",$timestamp),
+                        'updated' => 1
+                    );
+                    
+                    $result = $this->question_model->update_question($question_id, $data);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 200);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters are empty',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
+                    ], 401);
                 }
-
-                $timestamp = time();
-                // 2022-12-27 11:12:02
-
-                $data = array(
-                    'title' => $title,
-                    'body' => $body,
-                    'updatedAt' => date("Y-m-d h:m:s",$timestamp),
-                    'updated' => 1
-                );
-                
-                $result = $this->question_model->update_question($question_id, $data);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 200);
             } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -172,15 +185,28 @@
         // delete question
         public function delete_delete($question_id) {
            try {
-                if ($question_id === null || $question_id === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                $decodedToken = (array)decodeJWTToken($token);
+                $user_id = $this->question_model->get_user($question_id);
+                // only created user can delete the question
+                if ($isValid && $this->session->userdata('logged_in') && $user_id === $decodedToken['user_id']) {
+                    if ($question_id === null || $question_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters missing',
+                            'result' => null
+                        ], 400);
+                    }
+                    $result = $this->question_model->delete_question($question_id);
+                    $this->response([], 204);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters missing',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
+                    ], 401);
                 }
-                $result = $this->question_model->delete_question($question_id);
-                $this->response([], 204);
            } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -194,19 +220,30 @@
         // increase votes
         public function upvote_put($question_id) {
             try {
-                if($question_id === null || $question_id === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                // $decodedToken = (array)decodeJWTToken($token);
+                if ($isValid && $this->session->userdata('logged_in')) {
+                    if($question_id === null || $question_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters are missing',
+                            'result' => null
+                        ], 400);
+                    }
+                    $result = $this->question_model->increase_votes($question_id);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 200);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters are missing',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
+                    ], 401);
                 }
-                $result = $this->question_model->increase_votes($question_id);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 200);
             } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
@@ -221,19 +258,29 @@
         // decrease votes
         public function downvote_put($question_id) {
             try {
-                if($question_id === null || $question_id === "" ) {
+                $token = $this->input->request_headers()['x-auth'];
+                $isValid = validateJWTToken($token);
+                if ($isValid && $this->session->userdata('logged_in')) {
+                    if($question_id === null || $question_id === "" ) {
+                        $this->response([
+                            'code' => -1,
+                            'error' => 'mandatory parameters missing',
+                            'result' => null
+                        ], 400);
+                    }
+                    $result = $this->question_model->decrease_votes($question_id);
+                    $this->response([
+                        'code' => 0,
+                        'error' => null,
+                        'result' => $result
+                    ], 200);
+                } else {
                     $this->response([
                         'code' => -1,
-                        'error' => 'mandatory parameters missing',
+                        'error' => 'unauthorized',
                         'result' => null
-                    ], 400);
+                    ], 401);
                 }
-                $result = $this->question_model->decrease_votes($question_id);
-                $this->response([
-                    'code' => 0,
-                    'error' => null,
-                    'result' => $result
-                ], 200);
             } catch(Exception $e) {
                 log_message('error', 'error =>'.$e->getMessage());
                 $this->response([
